+++ examples/with-import/main.tf
@@ -47,7 +47,9 @@ module "essential_contacts" {
     {
       parent = "organizations/123456789"
       essential_contacts = {
-        "security@example.com" = ["SECURITY"]
+        "security@example.com" = {
+          notification_category_subscriptions = ["SECURITY"]
+        }
       }
     },
   ]