+++ examples/with-import/versions.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = ">= 4.62"
+      version = ">= 7.33"
     }
   }
 }